backend/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Config and validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Optional: load .env in development
python-dotenv>=1.0.0

# Structured logging (JSON, context)
structlog>=24.0.0

# MongoDB async driver
motor>=3.6.0
pymongo>=4.0.0

# Auth & security
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
